chore: switch benchmark to TypeScript
#149
